CLN: Remove redundant float() calls

+++ pandas/plotting/_matplotlib/converter.py
@@ -116,7 +116,7 @@ def deregister():
 
 
 def _to_ordinalf(tm: pydt.time) -> float:
-    tot_sec = tm.hour * 3600 + tm.minute * 60 + tm.second + float(tm.microsecond / 1e6)
+    tot_sec = tm.hour * 3600 + tm.minute * 60 + tm.second + tm.microsecond / 1e6
     return tot_sec
